# Optimize parser a bit